Side Kick
Every time the ball goes out of bonds in a side kick area, the ball is frozen in place and a side kicking area is spawned in the field, once you click a character, it's moved to the side kicking area.

Still needs to give the player the ability to choose where the character will go.